Clean up unnecessary casts and instance checks.

+++ src/main/java/micdoodle8/mods/galacticraft/core/client/gui/element/GuiElementSlider.java
@@ -162,7 +162,7 @@ public class GuiElementSlider extends GuiButton
 
     public double getColorValueD()
     {
-        return (this.sliderPos * 255.0D) / (double) (this.height - 1);
+        return (this.sliderPos * 255.0D) / (this.height - 1);
     }
 
     public int getButtonHeight()
